Increase pixel depth of vertex and normal maps
The vertex and normal map generation kernels now pack the data of output
images at 128 bits per pixel (32 bit, four channels). This is to allow a
wider range of values for each component of the verticies and to enable
single precision floating point values to be written into the normal map.